// react-frontend/src/api/models/menu-template-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spys API
 * REST API for Spys
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface MenuTemplateDTO
 */
export interface MenuTemplateDTO {
    /**
     * 
     * @type {number}
     * @memberof MenuTemplateDTO
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof MenuTemplateDTO
     */
    'description'?: string;
    /**
     * 
     * @type {number}
     * @memberof MenuTemplateDTO
     */
    'deliveryOffsetMinutes'?: number;
    /**
     * 
     * @type {number}
     * @memberof MenuTemplateDTO
     */
    'releaseOffsetMinutes'?: number;
    /**
     * 
     * @type {number}
     * @memberof MenuTemplateDTO
     */
    'orderByOffsetMinutes'?: number;
    /**
     * 
     * @type {string}
     * @memberof MenuTemplateDTO
     */
    'presetName'?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuTemplateDTO
     */
    'createdAt'?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuTemplateDTO
     */
    'updatedAt'?: string;
}